moving some more stuff around
