+++ src/editor/console/input/components/console-input.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "react-emotion";
 
-import DoubleChevron from "./double-chevron-icon";
-import ConsoleGutter from "../../history/components/console-gutter";
+import DoubleChevron from "../../double-chevron-icon";
+import ConsoleGutter from "../../console-gutter";
 import BaseIcon from "../../base-icon";
 
 import ConsoleInputField from "./console-input-field";
